Update tasks CRUD

+++ client/src/store/slices/tasksSlice.ts
@@ -1,5 +1,5 @@
 import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
-import { getHttpTasks, deleteHttpTask } from "../../api";
+import { getHttpTasks, deleteHttpTask, createHttpTask } from "../../api";
 import { ITasksState } from "../../types";
 
 export const getTasksThunk = createAsyncThunk(
@@ -16,12 +16,31 @@ export const getTasksThunk = createAsyncThunk(
   }
 );
 
+export const addNewTaskThunk = createAsyncThunk(
+  "tasks/post",
+  async(payload, { rejectWithValue }) => {
+    try {
+      // await createHttpTask(payload)
+      console.log("asdadasd")
+    } catch (error) {
+      const { message }: any = error;
+      console.log("error =>", message);
+      return rejectWithValue(message);
+    }
+  }
+)
+
 export const deleteTaskThunk = createAsyncThunk(
   "tasks/delete",
   async (taskId: number, { rejectWithValue }) => {
     try {
-      await deleteHttpTask(taskId);
-      return taskId;
+      const isDeleteTask = window.confirm("Delete task?")
+      if(!isDeleteTask) {
+        return
+      } else {
+        await deleteHttpTask(taskId);
+        return taskId;
+      }
     } catch (error) {
       return rejectWithValue(error);
     }
@@ -32,12 +51,18 @@ const initialState: ITasksState = {
   tasks: [],
   isFetching: false,
   error: null,
+  isEmpty: false,
+  isOpenNewTask: false
 };
 
 const tasksSlice = createSlice({
   name: "tasks",
   initialState,
-  reducers: {},
+  reducers: {
+    isOpenNewTaskWindow: (state) => {
+      state.isOpenNewTask = state.isOpenNewTask ? false : true;
+    }
+  },
   extraReducers: (builder) => {
     builder
       .addCase(getTasksThunk.pending, (state) => {
@@ -46,6 +71,7 @@ const tasksSlice = createSlice({
       })
       .addCase(getTasksThunk.fulfilled, (state, { payload }) => {
         state.isFetching = false;
+        state.isEmpty = state.tasks.length === 0 ? true : false;
         state.tasks = payload;
       })
       .addCase(getTasksThunk.rejected, (state, { payload }) => {
@@ -59,6 +85,9 @@ const tasksSlice = createSlice({
       })
       .addCase(deleteTaskThunk.fulfilled, (state, { payload }) => {
         state.isFetching = false
+        if(payload === undefined) {
+          return
+        }
         const findDeleteIndex = state.tasks.findIndex(task => {
           return task.id === Number(payload)
         })
@@ -68,9 +97,23 @@ const tasksSlice = createSlice({
         state.isFetching = false
         state.error = payload
       });
+    builder
+      .addCase(addNewTaskThunk.pending, (state) => {
+        state.isFetching = true;
+        state.error = null;
+      })
+      .addCase(addNewTaskThunk.fulfilled, (state, { payload }) => {
+        console.log("asdasdasdasdas")
+      })
+      .addCase(addNewTaskThunk.rejected, (state, {payload}) => {
+        state.isFetching = false;
+        state.error = payload;
+      })
   },
 });
 
 const { reducer } = tasksSlice;
+
+export const { isOpenNewTaskWindow } = tasksSlice.actions
 
 export default reducer;
